backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - MySQL
aiomysql==0.2.0
pymysql==1.1.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security and encryption
cryptography==41.0.7
bcrypt==4.1.2
pyjwt==2.8.0

# Data processing and analysis
sqlglot==27.0.0

# AI and OpenAI
openai==1.3.7
aiohttp==3.9.1
google-generativeai==0.3.2

# WebSocket support
websockets==12.0

# Caching
diskcache==5.6.3

# HTTP client
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# SQLite for metadata storage
aiosqlite==0.19.0
